backend/src/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
marshmallow>=3.0,<4
pymssql>=2.2
python-dotenv>=0.21

Flask-RESTX>=1.1.0
flask-swagger-ui
apispec
apispec_webframeworks

PyJWT>=2.8
PyYAML>=6.0
